Much improved

+++ cargo/R/onLoad.R
@@ -1,23 +1,32 @@
 .onLoad <- function(libname, pkgname) {
-  cache_dir <- tools::R_user_dir("cargo", "cache")
+  cache_dir <- cache_dir()
   if ( ! dir.exists(cache_dir) ) return()
   # Periodically delete cached material
-  last_purge_file <- file.path(cache_dir,"last-purge")
-  if ( ! file.exists(last_purge_file) ) return()
-  details <- readLines(last_purge_file)
+  last_purge_filename <- last_purge_filename()
+  if ( ! file.exists(last_purge_filename) ) return()
+  details <- readLines(last_purge_filename)
   if ( identical(details[1], "1") ) { # Details specification version 1
     names(details) <- c("specification", "date", "days_until_next")
     days <- as.integer(details['days_until_next'])
     if ( days > 0 ) {
       next_purge_threshold <- as.Date(details['date']) + days
-      if ( Sys.Date() >= next_purge_threshold ) purge_cache(cache_dir, days)
+      if ( Sys.Date() >= next_purge_threshold ) purge_cache(FALSE)
     }
+  } else {
+    unlink(cache_dir,  recursive=TRUE, force=TRUE)
   }
 }
 
-purge_cache <- function(cache_dir, days) {
-  unlink(file.path(cache_dir,"cargo","git"), recursive=TRUE, expand=FALSE)
-  unlink(file.path(cache_dir,"cargo","registry"), recursive=TRUE, expand=FALSE)
-  unlink(file.path(cache_dir,"rust_fn"), recursive=TRUE, expand=FALSE)
-  writeLines(c("1",as.character(Sys.Date()),days), file.path(cache_dir,"last-purge"))
+purge_cache <- function(last_purge_filename_only) {
+  cache_dir <- cache_dir()
+  if ( ! last_purge_filename_only ) {
+    unlink(file.path(cache_dir,"cargo","git"), recursive=TRUE, expand=FALSE)
+    unlink(file.path(cache_dir,"cargo","registry"), recursive=TRUE, expand=FALSE)
+    unlink(file.path(cache_dir,"rust_fn"), recursive=TRUE, expand=FALSE)
+  }
+  writeLines(c("1", as.character(Sys.Date()), days_until_next_purge), last_purge_filename())
 }
+
+days_until_next_purge <- 91
+cache_dir <- function() tools::R_user_dir("cargo", "cache")
+last_purge_filename <- function() file.path(cache_dir(), "last-purge")
